+++ CHANGELOG.md
@@ -24,6 +24,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New PBPB encoding for Z3 Solver
 - New UDPB encoding for Z3 Solver
 - `get_M()` function to handle M-variable access regardless of step order.
+- Metadata with detailed metrics information based on each instance solution
+- Graph visualizations on information of each instance.
 
 ### Changed
 - Segregated out and specified our current implementation of ORTools as UDPB (User-Dependent Pseudo-Boolean).